Add preview provider for ODF that uses Collabora Online's web API
This requires a Nextcloud that includes
https://github.com/nextcloud/server/pull/10275 .

To use, add OCA\Richdocuments\Preview\OpenDocument to an
'enabledPreviewProviders' array in the Nextcloud config/config.php.
Note that if you have 'enabledPreviewProviders' in config.php, you
need to explicitly include the built-in image format previewers there
that otherwise are enabled by default.

For example, like this:

'enabledPreviewProviders' =>
  array (
    'OC\Preview\PNG',
    'OC\Preview\JPEG',
    'OCA\Richdocuments\Preview\OpenDocument',
  ),